Qt5-Everywhere: show network error message
Instead of showing broken applications when there is no available
network, we now show an error message indicating that there is no
network available.  This is neat because now we dont have to let
customers know why things are working at shows, and people trying this
for themselves will know what they need to do to remedy the issue.

Change-Id: Idda904946fb823abce3f8f11c51b5516cb58472c
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@digia.com>