webengine: Add slight delay before loading the start page
This is a workaround that seems to prevent the random start-up
freeze on eAndroid.
The freeze is a deadlock in the graphics stack, specifically
in eglMakeCurrent, related to known GPU driver problems with
context sharing on Android on some devices. Chromium has some
workatounds for these problems, but our snapshot is not recent
enough to have reliable quiks for all configurations we support.
Since the QML Loader element used in qtlauncher and WebGL content
seem to be the only way to sporadically reproduce the freeze
this workaround should be sufficient until we do more testing
with an updated snapshot.

Change-Id: I02ec9f441c06ee3243e9975e4a49217400df27f2
Reviewed-by: Laszlo Agocs <laszlo.agocs@digia.com>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@digia.com>