Support Flat style and add new Controls demo
The Flat style for Qt Quick Controls is now enabled by
defualt for the launcher and there are a few places that
needed to be tweeked for use the correct style.  In addition
there is a new controls demo that shows off both the Enterprise
Controls from before along side the existing controls, but both
using the new Flat style which is only available to customers
and is intended for touch devices.

So we add one new demo, and disable 2 others (via exclude.txt)

Change-Id: I8aa6ddd7d48375fdf64150b0c5e2d451da07d00d
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@theqtcompany.com>