Add enterprise-kinectdatavis in tradeshow demos
This is a small demo I did with feedback from the sales guys in
Berlin. It's similar to the Kinect demo we had for 3D charts, except
that it doesn't require complicated packages like OpenNI; only
libfreenect to get the depth image.

I only tested the demo on the imx6. The mesh processing of incoming
frames is quite heavy so it might not work well on slower devices.
The resolution can be tweaked with QQuickFreenectDepthDataProxy's
width and height variables to change the resolution of the graph.

It isn't in basicsuits since it require a Kinect to be of any value.

Change-Id: I4b4619599842119e7d3e8ee130ece7244c0c645c
Reviewed-by: Michal Klocek <michal.klocek@digia.com>