Add physical screen size override slider to settings
The slider is in inches and assumes the usual 16:9 aspect ratio.

It will then rewrite appcontroller.conf to include QT_QPA_EGLFS_PHYSICAL_WIDTH
and HEIGHT. It can also parse the already set values.

This is essential because many framebuffer drivers do not provide any physical
size information, so overriding with the environment variables is often necessary.

Task-number: QTEE-935
Change-Id: I29452bc20cc58a259191fd65e7463f98ff792523
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@theqtcompany.com>