iot-sensortag: Add option to enable deploying resources to the file system
Add a config flag to the project file that controls how resources are
deployed with the binary. By default they are included in the binary
but it is also possible to deploy them to the file system.  This option is
needed as Qt5.7.1 and and Qt5.8 fail to load QML files from resource
files on some platforms (e.g. RPi3).

Change-Id: I7e1e938ac829d01ac6117e04bbaef053daf15154
Reviewed-by: Titta Heikkala <titta.heikkala@qt.io>