democompositor: Introduce an AppListModel to expose apps
Turn the AppEntry into a gadget and store the filename of where the
entry came from. Update the existing test to deal with it.

Introduce AppListModel as a QAbstractItemModel and expose the raw
AppEntry and its fields as role. Have some slow O(n) code to check
if an entry has already been added. The goal is to use a directory
watcher to automatically update this list.

Store the AppEntry as a pointer so we don't have to remove the
const from the field.

The next steps are to fully populate the AppListModel by scanning
and watching a directory and then replacing the LaunchButton code
with a Repeater on the application model. After starting of apps
will be done by passing the AppEntry to a launcher. Application
state should be stored outside the AppEntry and maybe in a different
model as well.

Change-Id: Ia933a09942ea8bb88eae5c101d13a30564d8c639
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>