democompositor: Expose the AppListModel and use to display apps
Replace the static configuration of the launch buttons with a Repeater
and the instantiated appModel. Use Component.onCompleted to parse the
app files. Maybe we can make this a property to ease starting it?

It is hardcoding the install directory for the *.app files but I am
not sure how to do it differently.

Change-Id: I5ef1d78a472f6e0f1dc1c0a3b050179949e6d818
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>