democompositor: Give up on having AppEntry with only const members
When building gadgets that include instances of AppEntry that will
expose these member variables a Q_PROPERTY we will end trying to do
a std::move (even if the AppEntry is not movable). Remove the const
from the member variables, leave the CONSTANT in the property to
not generate a write function and let's keep the discipline to not
assign to an existing AppEntry.

Change-Id: I742a9d091e7e6eac6cc3cb84a48b7b8812a24247
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>