democompositor: Provide an example to detect unknown apps
When a new wlShellSurface is created we can check the identity
of the client and if it is belongs to an application launched
by the democompositor. In the future such apps might be killed
and their surface not displayed.

Introduce appStateForPid to find the AppState and return it as
a QVariant to easily use it in QML. Use int as type as one can
not easily use Q_PID in QML.

Change-Id: Ibb9ac004a8016bd76f61679f5c837c99783fa7e8
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>