democompositor: Forward client application output
In the long run we could do line based reading and associate the
line with the AppState/AppEntry but for now it is better to see
some log messages at all (I was debugging an issue that with
QT_WAYLAND_DISABLE_WINDOWDECORATION set the surface would remain
black). Log output didn't help in the end but it was a start and
the issue is still open.

Change-Id: Icd4e041a4004c6a76e685906d63704c2cbf51f63
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>