democompositor: Remove deleted files from the model
In the previous code review Paul discovered the fact that files
would never be removed. Implement finding deleted files and emit
a signal per removed file. Consume the signal in QML and use the
new API to kill the process (one could close the wayland display
connection as well).

Emit the removed signals _after_ the model has been updated to
provide the QML code with an updated model during the signal
handling.

Change-Id: Ib3bb1bd953bfbf09d25c5e5822756a53b08fd797
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>