processlauncher: Fix crashing on destruction with running apps
The democompositor would crash because the QProcess will be deleted
from within ~QObject and then emit a signal which will access the
partially destructed WaylandProcessLauncher. Iterate through the
list and disconnect the slots and then delete the QProcess before
the ~QObject is ran.

An alternative would be to iterate the m_appStates and just block
the signals of the QProcess. This might be more future proof (e.g.
another signal being connected) but can impact other signal/slot
connections as well.

Change-Id: Id6c19dbe0c3070a131b28365584f06fb7c6014b3
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>